Added support spawntime in sql output support for 7.0.3

+++ WowPacketParserModule.V6_0_2_19033/Parsers/UpdateHandler.cs
@@ -96,7 +96,7 @@ namespace WowPacketParserModule.V6_0_2_19033.Parsers
             obj.Zone = CoreParsers.WorldStateHandler.CurrentZoneId;
             obj.PhaseMask = (uint)CoreParsers.MovementHandler.CurrentPhaseMask;
             obj.Phases = new HashSet<ushort>(V6_0_2_19033.Parsers.MovementHandler.ActivePhases);
-            obj.DifficultyID = (uint)CoreParsers.MovementHandler.CurrentDifficultyID;
+            obj.DifficultyID = CoreParsers.MovementHandler.CurrentDifficultyID;
 
             // If this is the second time we see the same object (same guid,
             // same position) update its phasemask
